Implement our own miniature build cache for module-based packages.
For packages that come from a module, go/build returns PkgObj path
inside the module root. Using it as-is leads to mixing source code and
binary artifacts and version control mishaps.

We solve the problem by detecting such paths and storing them in a
separate build cache directory. Note that our implementation is much
simpler and not compatible with the Go build cache. We rely upon
go/build returning unique PkgObj paths for different packages and simply
hash the original path to produce a path inside the cache.