Replace gotool package with `go list` and buildutil.ExpandPatterns()
gotool package matching logic was based on the old version of go (most
likely 1.8 or so), and its patterns were not matching packages from
third-party modules. We use `go list` instead to match packages on the
physical file system.

We also provide a shim for VFS-based build contexts using buildutil
package, which should be good enough for the few cases we have.