Support Go Modules in the `gopherjs serve` subcommand.
Under Go Modules package sources a no longer under a single GOPATH and
import paths can't be mapped onto the file system naively. Instead we
have to import the correct package and open the file relative to package
root.

Since we don't know which part of the requested path represents the
package import path, and which is a file within the package, we have to
make a series of guesses until one succeeds.