Add tests for GOPATH and Go Modules build modes.
We use https://github.com/gopherjs/todomvc as a simple test case for
either mode, since it is small and easy to debug, but still has a few
external dependencies to exercise Modules infrastructure.

I also added a step that verifies that the output in both modes is
identical. Strictly speaking this is not a requirement, but the current
implementation maintains this invariant, so we might as well test it.