Run go repo tests in parallel even in verbose mode.
This should significantly accelerate our CI runs, at a cost of a minor
risk of mixing up outputs from several independent tests. However, the
output is structured such that it should not be difficult to tell what's
going on and concurrency is low enough to make such conflicts pretty
unlikely anyway.