gopherjs test: execute tests for different packages in parallel.
This effectively implements a limited analog of go tool's -p flag, which
controls parallelism in the build toolchain. GopherJS's build session is
not thread-safe, so we still do all of that in the main goroutine, but
we can parallelize execution of each individual package tests.

The main objective of this change is to accelerate standard library test
execution in the CI, which currently runs single-threaded on a machine
with two cores.

When running with -p > 1, test output is redirected into a temporary
buffer to avoid mixing outputs of different package tests.

Note that this change does not affect parallelism within a single
package, these tests are still sequential unless t.Parallel() is used.