Support go1.17 slice to array pointer conversion.
Conversion is fully supported for slices of numeric types, which is
probably the most common use case for this feature judging by the
discussion in the proposal.

However, it is only partially supported for slices of complex types
(e.g. strings), since they are backed by the JavaScript's built-in Array
type, which lacks ability to share backing memory among subarrays.
Conversion works in cases when the slice is converted into array that
exactly matches the slice's underlying array, but panics if we are
trying to convert a subslice. I feel like an explicit failure is better
than a chance of introducing subtle bugs.

It also seems that GopherJS internally doesn't really distinguish
between array types and pointer to array types, which makes the whole
implementation somewhat messy when it comes to nil vs non-nil pointers
to arrays.

Last but not least, I've moved pointer cache for numeric arrays into the
backing ArrayBuffer, which ensures that pointers are comparable between
different subarrays.