Preserve nil slices when converting between different slice types.
Prior to this change the following would evaluate to false:

```
type mySlice []byte
_ = mySlice([]byte(nil)) == nil // Got: false, want: true
```

The fix is to check during conversion whether a slice is a nil slice and
return nil slice of the desired type in that case. GopherJS uses
sentinel values for nil slices, and technically they are are just
regular zero-length slices.