Provide correct constructor for names array-pointer types.
...and also document what "wrapped" types are about, because it took me
a looong time to figure out. Turns out this is pretty much a form of
primitive type boxing that is done in languages like Java.

Array-pointer types are one of the wrapped types, which means they are
represented by a native JS type at runtime, except for a few situations
when access to the full Go type needs to be preserved. To wrap such a
value, the native value is passed as a single constructor argument, but
named array pointer types did not provide that, using a default
constructor for all pointer types.