Add GitHub action workflows to track GopherJS output size changes.
`measure-size` is triggered by a pull request or an update to a pull
request. It compiles a reference app using GopherJS from a pull request
and from the target branch, and generates a report with output size
changes, if any.

`publish-size` is run upon completion of `measure-size` and posts the
generated comment as a comment in the pull requests. If there are
previous reports in the thread already, it will hide them to reduce
confusion.

The two steps are implemented as separate workflows to ensure that
untrusted code from a PR doesn't get executed in a workflow with a write
access to the repository. See [this
article](https://securitylab.github.com/research/github-actions-preventing-pwn-requests/)
for details.

Workflows use custom actions from
https://github.com/gopherjs/output-size-action.