Prevent event loop starvation by always scheduled goroutines.
If there are two or more goroutines that are constantly interacting with
each other, they both will regularly give back control to the scheduler.
But if scheduler never yields to the JavaScript's event loop, events
like IO or timers will never have a chance to fire, and, for example,
`time.Sleep()` will never return.

We can't yield to the event loop after every goroutine switch, the
overhead of `setTimeout()` is ~1000x higher than the tight while-true
loop.

To amortize the overhead we will how yield if there's no scheduled
goroutine (old behavior), or if we spent more than 4ms executing the
code. In my benchmark, the cost of goroutine switch changed from
273ns/op to 341ns/op, which we can live with — most apps don't have such
tight loops anyway.