Correctly handle stack unwinding after a panic in a deferred function.
Runtime throws null when it need to unwind the stack from a point where
`panic` was called to where panic was handled. However, if the panic was
thrown in a deferred function and recovered in the subsequent deferred
function of the same call stack frame, this throw went unhandled
crashing the program.

This change adds a point to handle such throws giving a chance to either
stop unwinding or continue invoking deferred functions if another panic
happened.

Fixes https://github.com/gopherjs/gopherjs/issues/780.