Correctly handle built-ins and `js.Object` methods with `go` keyword.
Built-ins and `js.Object` methods are represented by JS expressions that
may not be straightforward function calls, or reference functions
defined in the prelude. In such cases, the callable expression must be
wrapped in a lambda.

https://github.com/gopherjs/gopherjs/pull/998 has fixed a similar issue
for the `defer` keyword, and this change generalizes the same fix for
the `go` keyword.

Fixes https://github.com/gopherjs/gopherjs/issues/547.