Remove GOROOT and GOPATH fields from `build.Options`.
They were used as a side channel to extract GOROOT and GOPATH from
`build.Session`, which is an incorrect use of options. Eventually all
logic that requires these paths should be moved to the `build` package,
but for now we just provide an additional method to create a source
mapping callback.