Make `internal/intern` unsupported and avoid its usage in `net/netip`.
The optimization this package implements relies heavily on finalizers,
which GopherJS doesn't support. As such it would create a memory leak
rather than an actual optimization. Until we decide that we can use
WeakMap API, this package can't be safely supported.

The only current use of this package is in the `net/netip` package and
this commit adds overlays that replace uses of the intern package with
direct use of strings.