Support .inc.js files for standard library overlays.
This will simplify access to low-level JavaScript features from the
overlay sources. In many ways, this will fill the role of assembly in
the upstream standard library.

This change also introduces stricter checks for the ad-hoc build mode,
requiring that all sources are in the same directory. This is necessary
to maintain invariants about the build.Package we construct and is
similar to what the go tool does.