runtime: Filter call stack returned by `Callers()` and `Caller()`.
This change makes the returned stack traces more similar to what you'd
get in a regular Go runtime. In particular, it skips certain
GopherJS-specific prelude functions like `$callDeferred()`. I expect we
will be adding to the list as we discover other edge cases.

We also rename certain functions to match their upstream counterparts,
since parts of the standard library rely on this. We mark them as
reserved symbols for prelude minification to make sure this doesn't
break when minified.