Change Go map implementation from object properties to ES2015 Map.
The main motivation for this change is to improve the performance of
len() on maps.  Previously, len() would compile to javascript that
called `Object.keys().length`.  This iterates the whole map to get its
length.

Using the benchmarks added in this commit, on a map with 10000 elements:

```
BenchmarkMapLen             	    9886	    115719 ns/op
```

Now, len() just calls to javascript Map API, Map.size, which now as fast as calling len() on a slice.

```
BenchmarkMapLen             	1000000000	         0.5100 ns/op
```

This pull request contains changes to the compiler, prelude and standard
library necessary to utilize the new API. However, when externalizing
maps, they are still converted into objects to maintain backwards
compatibility.