compiler/gopherjspkg: explicitly mention the vfsgen dependency
The natives package has a blank import of the vfsgen package so that
it stays included in go.mod (otherwise go mod tidy considers it as
unused and removes it). Copy that mechanism here.

Place it under a file with a "generate" build constraint so it doesn't
get included in a normal gopherjs build, since it's only needed during
code generation.