compiler: go:linkname support method (#1152)
compiler: support `go:linkname` directive for methods

This is similar to what the upstream compiler supports. This functionality is inherently unsafe, but can be useful for some certain libraries like [reflectx](https://github.com/goplusjs/reflectx/blob/main/name_js.go#L17). The first argument of the function will act as a receiver of the linked method.

As long as underlying typed for the first arguments match, they will be converted at runtime, which allows linking to methods of unexported types. However, types of the other arguments are not converted, nor the signature is verified to match the linked method.