Support calling methods on wrapped generic types.
Since we are not using monomorphization, the generated code must be able
to work with both wrapped and non-wrapped types. Here wrapped types are
the types like integers, which gopherjs represents at runtime using
native JS types like `number` rather than gopherjs-specific objects,
which improves execution efficiency. This also means that those types
lack Go-specific metadata and are _wrapped_ into gopherjs-specific
objects only when necessary, e.g. before calling a method. In many ways
this is similar to boxing/unboxing in other languages.

This presents a problem for generic code, though, since the types
derived from typeparams the compiler can't be sure that they are wrapped
or not. We handle this by adding a new `.wrap()` method to each type
object. For wrapped types it will wrap the raw value, and for regular
types it will be a no-op. Whenever we need a receiver for a type that
depends on a typeparam `T`, we will pass the value through `T.wrap()`.

This is a slightly bigger overhead for method calls compared to
non-generic code, but hopefully JIT should be able to inline it well
enough:

  - For each type the `wrap()` method is static throughout the lifetime
    of the program and is a pure arrow function.
  - The function has no branching: it either always returns the same
    value (for non-wrapped types) or always wraps it.

Although I'm not 100% confident this fix covers all edge cases, it seems
to work with embedded types and fixes 4 test cases from the typeparam
test suite.