compiler: factor out utility types for processing Go sources and errors.
This is the first step in reducing complexity of the compiler.Compile
function.

The new `sources` type represents all inputs into the package
compilation and simplifies extracting useful information out of them. It
is designed to have little business logic of its own and serves as a
convenient bridge to other packages like go/types or astrewrite.

The ErrorList type is extended with utility methods that reduce
verbosity of the calling code.