README: document the current GOOS and GOARCH values
GopherJS 1.18 has standardized on a single value for GOOS and GOARCH
in PR #1111. Update the README accordingly. (There has been agreement
in issue #1169 that it's okay to be updating the top-level README for
GopherJS 1.18-specific changes by now.)

Also reword advice in compatibility.md to be more general. The upstream
Go WebAssembly port (i.e., GOOS=js GOARCH=wasm, and notably not wasm64)
is limited to 4 GiB of linear memory, it just happens to at this time
use 64-bit variables for int/uint/uintptr types. So maybe it's not a
great example to refer to. But a reminder that some types vary in size
based on the architecture or implementation is probably worth including
anyway.