Additional nil checks around NodeJS APIs.
These changes make gopherjs work correctly in environments with fake
`$global.process` object, which may be missing some of the properties.
Specifically, when Go's `wasm_exec.js` is used, it fakes that object,
but only partially. So if a gopherjs program was loaded in the same
context as a Go wasm program, gopherjs would throw an undefined property
exception.