Fix infinite recursion when initializing generic types.
Prior to this change, when a generic type depended on itself,
instantiating it would cause infinite recursion, for example:

```go
type List[any T] struct {
  next *List
}
```

In its generic factory function, it would attempt instantiate `*List`
before `List`, since the latter depends on the former. However, for that
to succeed `List` must be already instantiated, but it wouldn't be, yet.

We break the dependency loop in the same way it's done for non-generic
types: we declare variables for the anonymous types upfront, so that
they could be captured by the type constructor closure, but only assign
them values after the type instance has been created and stored in
`$typeInstances`. In that case a call to `List(T)` would return the
cached instance, avoiding the recursion.