Fix implicit conversions bug in the assignment operator.
Go considers types T and V assignable to each other if both have the
same underlying type and at least one of them is not a named type.
However, those are still distinct types, for example for reflect
purposes.

Prior to this change gopherjs would simply assign the value of type T to
type V without converting its type, which is incorrect. This change
forces type conversion to happen in such cases in order to maintain the
correct runtime type.