Fix taking a pointer to array/slice element of a typeparam type.
Another case of pointer-to-struct being a special case in GopherJS. Both
a struct and a pointer to it are essentially represented by the same JS
object (due to the pass-by-reference semantics JS objects have), so
wrapping it into an extra pointer breaks method calls.

Because for type params we can't know whether the concrete type will be
a struct or not at compile time, we have to make the check in the
runtime.