Foundations of type conversion support for type params.
Since GopherJS doesn't generate instantiation-specific code for generic
functions and types, we have to do conversion at runtime. This commit
creates basic infrastructure for such conversion, and uses int64 and
uint64 types as a first demonstration of concept.

Each type `T` provides a `T.convertFrom` method, which accepts a value
of any compatible type `Q` and returns a corresponding value of type
`T`.

This conversion will be less efficient compared to non-generic code,
since we have to branch based on the input type every time we do the
conversion, but this is the price for more compact code.