Throw an error on conversion between a typeparam and unsafe.Pointer.
Unsafe pointers are not well supported by GopherJS and whatever support
exists is not well documented. Implementing this conversion will require
a great deal of great reverse engineering, and I'm not sure anyone will
ever need this.

So for now, attempting such a conversion will simply throw an
"unimplemented" exception.