Implement typeparam conversion to interface.
In general, this conversion is almost a no-op, except that previously
*js.Object instances were not considered a wrapped type (which is
incorrect) and were not properly wrapped when converting to an
interface. Other than that, GopherJS's representation for an interface
value is actually its concrete value.

Testing the conversion is a bit tricky, since we can't use reflect
because it relies on conversion to `any` interface being correct -
something we can't take for granted. To make testing possible, I've
refactored the tests a bit and use low-level JS manipulation to inspect
values that result from conversion.