Support append() and delete() built-ins for type params.
Instead of interrogating the type of the slice or map expression passed
to the built-in, we interrogate types of the computed built-in
signature. In case of append(), it will be a slice of element types due
to variadic nature of the built-in. For map it will be the plain key
type. We can then use that type to translate expressing with implicit
conversion.