Implement a new internal library for discovering generic instances.
The library scans type-checked Go AST and incrementally discovers
instantiations of generic types and functions in it. Each instantiation
is recorded with the origin object and a set of type arguments.

The scan is seeded from two sources:

 - Any Instances that have been added to the working InstanceSet
   externally (for example, from references by other packages).
 - Any instances that are discovered within the non-generic code in the
   package itself.

The discovery proceeds to search for instances in the generic code by
traversing its AST for each set of previously discovered type arguments.
In that scan, whenever a type is defined in terms of a type param, we
map it onto a concrete type given the current set of type arguments.

For now, the library is not fully ready to handle generic instances that
come across the package boundaries, that will come later.

The type mapping logic turns out to be fairly complex, so instead of
reimplementing it I decided to cheat and "borrow" it from the go/types
package via a go:linkname directive. I didn't want to vendor it because
I expect it would be changing fairly regularly in future. We shall see
whether this was a wise choice.

This commit also includes a couple of supporting refactorings:

 - Moved the symbols logic used by go:linkname to a separate package,
   since it turned out to be useful for generics as well.
 - Added a testingx helper library with a generic `Must(t)` helper to
   make trivial error handling in tests less verbose.