Type parameter substitution when translating function instances.
Instead of using type information directly from types.Info.TypeOf, we now
use a funcContext.typeOf, which passes the resulting type through
typeparams.Resolver. If we are translating a parameterized code, it will
perform type parameter substitution according to the type arguments of
the instance we are currently translating.

This allows us to generate a type-specific code for each function
instance, making sure correct operations are used in each case.