Implement generic type instantiation.
Similar to standalone functions, we generate a set of initialization
code for each combination of type parameters discovered. Code for
metadata initialization is generated with type argument substitution
active for the given instantiation. Methods are translated essentially
the same as standalone functions, except that they are assigned to the
receiver type object instead of a standalone variable.