Correctly handle all integer types in funcContext.formatExpr().
Previously passing any type other than int would cause a type assertion
failure and a panic. This bug could be triggered when evaluating a
result of len() built-in for array types. However, it remained hidden in
non-generic code, since the expression was determined to be constant,
short-circuiting translation. Apparently, in generic code go/types can't
prove the constant value, thus revealing the bug.