Discover generic type declarations inside generic functions.
Prior to this change, while scanning for seed generic instances we would
not traverser generic function (or method) bodies, which lead to generic
types declared inside them not being added to objMap. Later, when an
instance of such type was discovered, we had no matching AST node to
process for further instantiations.

With this change, we do traverse generic function bodies, but only to
populate objMap. We ignore potential type instantiations in this case
because they would be defined in terms of a type param, which are not
interesting to us.