Perform type substitution for method selection expressions.
When calling a method on a type param or a type defined using a type
param we replace the original selection from go/types.Info.Selections
with a synthetic counterpart for the instantiated receiver type.

Unfortunately, we can't construct instances of types.Selection due to
its fields being private, so we create an interface-compatible
implementation instead.