Fix variable name allocation for inline types.
Prior to generics an inline type was always first encountered by the
compiler at its decl statement, where it reserved the variable name for
it. With generics, it may be encountered before that when translating
one of the instantiations of a generic type.

With this change, variable name allocation will happen correctly
regardless of where it would be encountered first.