Fix variable name assignment for result variables.
Prior to this change, when a generic function had named result
variables, we had an identity mismatch when accessing result variables
through the instantiated function signature vs through the function body
AST. The latter references the types.Var instance defined in terms of
type parameters, whereas the former produced synthetic instantiated
versions of the same types.Var objects. As a result, two different names
were assigned to what should be the same object, depending on how it was
reached.

With this change, we preserve the original, uninstantiated signature,
which result variables match the ones referenced in the body AST.
Whenever we need access to the instantiated signature types we perform
type substitution ad-hoc.