Perform type substitution in the type switch expression.
Prior to this change, we did not substitute type parameters with
instantiated counterparts, which led to the incorrect treatment of the
type param types as interfaces and invalid type assertion results. This
change corrects the mistake.

Although typeswitch2.go and typeswitch5.go generate a different output
from upstream Go, it is semantically equivalent and the difference is
down to console.log() behaving slightly differently than println() in
native Go.