Borrow type substitution logic from x/tools/go/ssa package.
Previously we've been using similar functionality from go/types package.
Since it was unexported, we've used a //go:linkname directive to gain
access to it. However, upgrade to go1.19 changed arguments of the
function in ways that are difficult to replicate, so I switched to an
equivalent reimplementation from the go/ssa package. This version is
implemented in terms of public go/types interface, so it's easy to
vendor and update.

To simplify code updates in future I kept all original code intact and
added a separate source file that exports the necessary functions for
gopherjs use.