Fix float64 to uint{8,16,32} conversion.
Before this change, GopherJS compiler emitted `(f >> 0)` expression to
convert a float64 `f` to any non-64-bit unsigned integer type. This is
incorrect, because `>>` is a signed bitshift operator in JS, so the
returned value remained signed. Moreover, it did not take into account
to bit size of the target type.

By removing the switch cause, we make the compiler fall through to the
default clause where `fc.fixNumber()` actually does the right thing,
taking the target into account.

Fixes #733.