Include original function names into source maps.
The main change in this commit is an ability to use identifier name
mapping, which we use to show original function names in the source map.

This addresses the long-standing https://github.com/gopherjs/gopherjs/issues/1085,
where GopherJS call stacks were somewhat difficult to interpret due to
function name mangling, especially in minified form. Now we emit an
additional source map hit with the original function name, which Node is
able to pick up.

While at it, I moved source map hinting logic into its own package with
tests and added some documentation on how it works. Now it should be
easy to extend this mechanism for even richer source maps if we want to.